Moves Component-based Ble into file back up in readiness for transformation to hooks
